requirements.txt
azure-ai-textanalytics
pymongo
python-dotenv
